Version 25

+++ app/Services/LigneFraisFservice.php
@@ -10,9 +10,18 @@ use function Symfony\Component\String\s;
 
 class LigneFraisFservice
 {
-	public function ajouterUnFraisFpourUnFrais(LigneFraisF $ligne_fraisforfait) {
+	public function ajouterUnFraisFpourUnFrais(LigneFraisF $ligne_fraisforfait, $id_visiteur) {
         try {
-            $ligne_fraisforfait->save();
+            $visiteurDuFrais = Frais::query()
+                ->select('id_visiteur')
+            ->find($ligne_fraisforfait->id_frais);
+            if ($visiteurDuFrais->id_visiteur != $id_visiteur) { // Ne pas mettre !== ou === pour vérifier ces 2 valeurs, mais != ou == (là != du coup)
+                throw new UserException(
+                    "Accès refusé", "Tu ne peux pas ajouter de frais au forfait à ce frais"
+                );
+            } else {
+                $ligne_fraisforfait->save();
+            }
         } catch (QueryException $exception) {
             $userMessage = "Erreur d'accès à la base de données";
             throw new UserException(
@@ -23,13 +32,22 @@ class LigneFraisFservice
         }
     }
 
-	public function modifierUnFraisFpourUnFrais(LigneFraisF $ligne_fraisforfait) { // Solution pour que ça marche trouvée par ChatGPT, ça ne marchait pas avec save()
+	public function modifierUnFraisFpourUnFrais(LigneFraisF $ligne_fraisforfait, $id_visiteur) { // Solution pour que ça marche trouvée par ChatGPT, ça ne marchait pas avec save()
         try {
-            LigneFraisF::where('id_frais', $ligne_fraisforfait->id_frais)
-            ->where('id_fraisforfait', $ligne_fraisforfait->id_fraisforfait)
-            ->update([
-                'quantite_ligne' => $ligne_fraisforfait->quantite_ligne
-            ]);
+            $visiteurDuFrais = Frais::query()
+                ->select('id_visiteur')
+                ->find($ligne_fraisforfait->id_frais);
+            if ($visiteurDuFrais->id_visiteur != $id_visiteur) { // Ne pas mettre !== ou === pour vérifier ces 2 valeurs, mais != ou == (là != du coup)
+                throw new UserException(
+                    "Accès refusé", "Tu ne peux pas modifier le frais au forfait de ce frais"
+                );
+            } else {
+                LigneFraisF::where('id_frais', $ligne_fraisforfait->id_frais)
+                    ->where('id_fraisforfait', $ligne_fraisforfait->id_fraisforfait)
+                    ->update([
+                        'quantite_ligne' => $ligne_fraisforfait->quantite_ligne
+                    ]);
+            }
         } catch (QueryException $exception) {
             $userMessage = "Erreur d'accès à la base de données";
             throw new UserException(
@@ -40,27 +58,25 @@ class LigneFraisFservice
         }
     }
 
-	public function deleteFraisF($id_frais, $id_fraisF, $id_visiteur) {
+	public function deleteFraisFpourUnFrais($id_frais, $id_fraisF, $id_visiteur) {
         try {
-            $unFraisF = LigneFraisF::query()
-				->select()
-				->where('id_fraisforfait', '=', $id_fraisF)
-				->where('id_frais', '=', $id_frais);
-			$unFraisF->delete();
-				//$visiteurDuFrais = Frais::query() // Pour mettre une erreur si le frais du frais hors forfait n'est pas de notre compte
-				//	->select('id_visiteur')
-				//	->where('id_frais', '=', $id_frais);
-				//if ($visiteurDuFrais->id_visiteur != $id_visiteur) {
-				//	throw new UserException(
-				//		"Tu n'as pas accès à ce frais au forfait"
-				//	);
-				//} else {
-				//	$unFraisF->delete();
-				//}
+            $visiteurDuFrais = Frais::query()
+                ->select('id_visiteur')
+                ->find($id_frais);
+            if ($visiteurDuFrais->id_visiteur != $id_visiteur) { // Ne pas mettre !== ou === pour vérifier ces 2 valeurs, mais != ou == (là != du coup)
+                throw new UserException(
+                    "Accès refusé", "Tu n'as pas accès à ce frais au forfait"
+                );
+            } else {
+                $unFraisF = LigneFraisF::query()
+                    ->select()
+                    ->where('id_fraisforfait', '=', $id_fraisF)
+                    ->where('id_frais', '=', $id_frais);
+                $unFraisF->delete();
+            }
         } catch (QueryException $exception) {
             if ($exception->getCode() == 23000) {
                 Session::put('erreur', $exception->getMessage());
-                return redirect(url('/Frais/modifier/'.$id_frais.'/forfait/lister'));
             } else {
                 return view('error', compact('exception'));
             }
